♻️ Extract components (#731)
